Refactor ipopt C interface
This commit makes large improvements to the C interface for ipopt and so to the rust interface.
Notably, scaling was broken in a previous commit, it should be fixed now although it lacks tests.
Initialization and problem size specification is now done through callbacks, which is how it is done
on the C++ side. This reduces the boiler plate for the C interface, and enables the user to specify
initialization and size between subsequent calls to the same ipopt object. This commit also remove
much of the needless allocation done in the original C interface (probably because it needed to be
more general).