Revamp sys crate
This commit removes the C API hack used to replace Ipopt's own capi with
a custom one. Instead we build a separate library called CNLP that
exposes a custom C API and links against an Ipopt binary. This way we
have the flexibility of chosing different Ipopt installs without having
to build it ourselves.

Another addition in this commit is that of improved build process where
the ipopt library binaries can be downloaded from the JuliaOpt
organization. Although this is not foolproof since for instance the
users would require to have libgfortran installed. This brings us closer
to a more reliable build process.